seperated WebViewForDetailView to Views/Detail
