+++ PriceTracker/PriceTracker/ViewModels/DetailViewModel.swift
@@ -6,21 +6,7 @@
 //
 
 import SwiftUI
-import WebKit
 
-struct WebViewForDetailView: UIViewRepresentable {
-    var url: URL
-
-    func makeUIView(context: Context) -> WKWebView {
-        let webView = WKWebView()
-        return webView
-    }
-
-    func updateUIView(_ uiView: WKWebView, context: Context) {
-        let request = URLRequest(url: url)
-        uiView.load(request)
-    }
-}
 
 class DetailViewModel: ObservableObject {
     @Published var gameDetail: DetailModel?
